Backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.6
gunicorn==23.0.0
